Time: 454 ms (29.87%), Space: 76.6 MB (96.1%) - LeetHub

+++ 2191-sort-the-jumbled-numbers/2191-sort-the-jumbled-numbers.cpp
@@ -9,8 +9,8 @@ public:
 
     static int transform_int(std::vector<int>&mapping, int num) {
         if (num == 0) return mapping[0];
-        int64_t result = 0;
-        int64_t power = 1;
+        int result = 0;
+        int power = 1;
         do {
             result = power * mapping[num % 10] + result;
             power *= 10;
